Move `ui/language/LanguageUtilsTest.kt` into `model/common/user/UtilsTest.kt`
