+++ app/src/androidTest/java/com/github/se/signify/model/common/user/UtilsTest.kt
@@ -1,15 +1,14 @@
-package com.github.se.signify.ui.language
+package com.github.se.signify.model.common.user
 
 import android.content.Context
 import androidx.test.core.app.ApplicationProvider
 import com.github.se.signify.model.common.updateLanguage
-import java.util.Locale
-import org.junit.Assert.assertEquals
-import org.junit.Assert.assertThrows
+import org.junit.Assert
 import org.junit.Before
 import org.junit.Test
+import java.util.Locale
 
-class LanguageUtilsTest {
+class UtilsTest {
 
   private lateinit var context: Context
 
@@ -25,23 +24,24 @@ class LanguageUtilsTest {
     val targetLocale = Locale(targetLanguageCode)
 
     // Call the function to update the language
-    updateLanguage(context, targetLanguageCode)
+      updateLanguage(context, targetLanguageCode)
 
     // Verify that the app's locale has been updated
     val updatedLocale = context.resources.configuration.locales[0]
-    assertEquals(
-        "The app's locale should be updated to the target language", targetLocale, updatedLocale)
+      Assert.assertEquals(
+          "The app's locale should be updated to the target language", targetLocale, updatedLocale
+      )
 
     // Cleanup: Restore the initial locale
-    updateLanguage(context, initialLocale.language)
+      updateLanguage(context, initialLocale.language)
   }
 
   @Test
   fun updateLanguage_throwsErrorWhenInvalidLanguageCode() {
     val targetLanguageCode = "DE"
     // Call the function to update the language
-    assertThrows(IllegalArgumentException::class.java) {
-      updateLanguage(context, targetLanguageCode)
-    }
+      Assert.assertThrows(IllegalArgumentException::class.java) {
+          updateLanguage(context, targetLanguageCode)
+      }
   }
 }
