+++ app/src/test/java/com/github/se/signify/model/common/UtilsTest.kt
@@ -1,9 +1,6 @@
-package com.github.se.signify.model
+package com.github.se.signify.model.common
 
 import com.github.se.signify.R
-import com.github.se.signify.model.common.getIconResId
-import com.github.se.signify.model.common.getImageResId
-import com.github.se.signify.model.common.getTipResId
 import junit.framework.TestCase.assertEquals
 import org.junit.Test
