+++ app/src/test/java/com/github/se/signify/model/common/user/SaveUserToFireStoreTest.kt
@@ -1,8 +1,7 @@
-package com.github.se.signify.model.user
+package com.github.se.signify.model.common.user
 
 import android.os.Looper
 import androidx.test.core.app.ApplicationProvider
-import com.github.se.signify.model.common.user.saveUserToFireStore
 import com.google.android.gms.tasks.Task
 import com.google.android.gms.tasks.Tasks
 import com.google.firebase.FirebaseApp
