+++ app/src/test/java/com/github/se/signify/model/home/quiz/QuizRepositoryFirestoreTest.kt
@@ -1,7 +1,6 @@
-package com.github.se.signify.model.quiz
+package com.github.se.signify.model.home.quiz
 
 import androidx.test.core.app.ApplicationProvider
-import com.github.se.signify.model.home.quiz.QuizRepositoryFireStore
 import com.google.android.gms.tasks.OnCompleteListener
 import com.google.android.gms.tasks.Task
 import com.google.firebase.FirebaseApp
