+++ app/src/test/java/com/github/se/signify/model/profile/stats/MockStatsRepositoryTest.kt
@@ -1,7 +1,5 @@
-package com.github.se.signify.model.stats
+package com.github.se.signify.model.profile.stats
 
-import com.github.se.signify.model.profile.stats.MockStatsRepository
-import com.github.se.signify.model.profile.stats.Stats
 import org.junit.Assert.assertEquals
 import org.junit.Assert.assertFalse
 import org.junit.Assert.assertNull
