+++ app/src/test/java/com/github/se/signify/model/profile/stats/StatsViewModelTest.kt
@@ -1,10 +1,7 @@
-package com.github.se.signify.model.stats
+package com.github.se.signify.model.profile.stats
 
 import com.github.se.signify.model.authentication.UserSession
 import com.github.se.signify.model.dependencyInjection.MockDependencyProvider
-import com.github.se.signify.model.profile.stats.MockStatsRepository
-import com.github.se.signify.model.profile.stats.Stats
-import com.github.se.signify.model.profile.stats.StatsViewModel
 import org.junit.Assert.assertEquals
 import org.junit.Assert.assertTrue
 import org.junit.Before
